% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{SpaMTPdbBundle}
\alias{SpaMTPdbBundle}
\title{Retrieve a coherent set of SpaMTP annotation resources}
\usage{
SpaMTPdbBundle(
  resources = SpaMTPdbResources(version = "latest", default_only = TRUE)$resource,
  version = "latest",
  ...
)
}
\arguments{
\item{resources}{Character vector of resource names. By default, all core
resources used by SpaMTP are returned.}

\item{version}{Resource version or \code{"latest"}.}

\item{...}{Passed to \code{\link[=SpaMTPdbResource]{SpaMTPdbResource()}}.}
}
\value{
A named list of resources from one database version.
}
\description{
Retrieve a coherent set of SpaMTP annotation resources
}
\examples{
SpaMTPdbBundle(resources = "chem_props", metadata = TRUE)
}
