% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{SpaMTPdbResources}
\alias{SpaMTPdbResources}
\title{List SpaMTPdb resources}
\usage{
SpaMTPdbResources(version = NULL, category = NULL, default_only = FALSE)
}
\arguments{
\item{version}{RaMP/SpaMTPdb resource version. \code{NULL} selects all versions.}

\item{category}{Optional resource category.}

\item{default_only}{Return only resources used by the default SpaMTP
annotation and pathway pipelines.}
}
\value{
A data frame describing available resources.
}
\description{
List SpaMTPdb resources
}
\examples{
SpaMTPdbResources(default_only = TRUE)
}
