% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{spaMTPdbVersion}
\alias{spaMTPdbVersion}
\title{Report the available SpaMTPdb versions}
\usage{
spaMTPdbVersion()
}
\value{
A character vector of resource versions, newest first.
}
\description{
Report the available SpaMTPdb versions
}
\examples{
spaMTPdbVersion()
}
