% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaMTPdb-package.R
\docType{package}
\name{SpaMTPdb-package}
\alias{SpaMTPdb}
\alias{SpaMTPdb-package}
\title{SpaMTPdb: annotation resources for SpaMTP}
\description{
\code{SpaMTPdb} provides version-aware access to RaMP-derived chemical,
identifier, pathway, and pathway-network resources. Large resources are
retrieved through AnnotationHub; package developers may instead configure a
directory containing staged \code{.rds} files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BCRL-tylu/SpaMTPdb}
  \item Report bugs at \url{https://github.com/BCRL-tylu/SpaMTPdb/issues}
}

}
\author{
\strong{Maintainer}: Tianyao Lu \email{lu.t@wehi.edu.au}

Authors:
\itemize{
  \item Tianyao Lu \email{lu.t@wehi.edu.au}
  \item Andrew Causer
  \item Quan Nguyen
}

}
\keyword{internal}
