% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{SpaMTPdbVersion}
\alias{SpaMTPdbVersion}
\title{Report the available SpaMTPdb versions}
\usage{
SpaMTPdbVersion()
}
\value{
A character vector of resource versions, newest first.
}
\description{
Report the available SpaMTPdb versions
}
\examples{
SpaMTPdbVersion()
}
